requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
PyYAML>=6.0
plotly>=5.17.0
boto3>=1.28.0
numpy>=1.24.0
python-dotenv>=1.0.0
